v2.1 1、加入spring框架，更改用例xml格式，实现spring注入bean

+++ src/com/interf/eyee/utils/assertutils/BaseAssertUtil.java
@@ -9,7 +9,7 @@ import com.interf.eyee.entity.BaseDataEntity;
  *
  */
 public interface BaseAssertUtil {
-	public abstract void assertCode(BigInteger actual, BigInteger expect);
+	public abstract void assertCode(long actual, long expect);
 	public abstract void assertMsg(String actual, String expect);
 	public abstract <T extends BaseDataEntity> void assertData(T actual, T expect);
 }
